Changed user interface to determine confidence intervals for SmoothedPG (i. e., integrated getCoherencePointwiseCIs and get CoherencyPointwiseCIs into getPointwiseCIs).
Added parameter to getValues-QuantilePG to allow for the selection of
dimensions.

Fixed plot-QuantilePG so that only the first component is plotted.

Fixed a bug in getValues-QuantileSD, getCoherency-QuantileSD and
getCoherency-SmoothedPG, where only the result was returned as an array
with the wrong dimensions.

Updated ref-results.rdata so that the test regarding standard errors can
be run again.